+++ jsconfig.json
@@ -1,8 +1,9 @@
 {
   "compilerOptions": {
+    "jsx": "react",
     "baseUrl": "src/",
     "paths": {
-      "@/*": ["./*"]
+      "@/*": ["./src/*"]
     }
   }
 }
