+++ package.json
@@ -11,17 +11,26 @@
   },
   "dependencies": {
     "@headlessui/react": "^1.4.2",
+    "@hookform/resolvers": "^4.1.3",
+    "@radix-ui/themes": "^3.2.1",
     "axios": "^0.21.1",
     "next": "^14.2.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "swr": "^1.3.0"
+    "react-hook-form": "^7.54.2",
+    "react-query": "^3.39.3",
+    "swr": "^1.3.0",
+    "workbox-window": "^7.3.0",
+    "zod": "^3.24.2",
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.12.1",
     "@babel/preset-react": "^7.18.6",
     "@next/eslint-plugin-next": "^14.2.3",
     "@tailwindcss/forms": "^0.5.2",
+    "@types/node": "22.13.10",
+    "@types/react": "19.0.12",
     "autoprefixer": "^10.4.2",
     "eslint": "^8.57.1",
     "eslint-config-next": "^14.2.24",
