+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import LoginLinks from "@/app/LoginLinks";
 import ApplicationLogo from "@/components/custom-icons/ApplicationLogo";
 import Link from "next/link";
+import LoginLinks from "@/app/LoginLinks";
 
 export const metadata = {
   title: "Unos y Otros",
@@ -9,26 +9,24 @@ export const metadata = {
 const Home = () => {
   return (
     <>
-      <div className="container mx-auto max-w-6xl px-4 py-4 flex justify-between items-center">
+      <div className="container mx-auto max-w-6xl px-[24px] py-[16px] flex justify-between items-center">
         <div className="flex items-center">
           <ApplicationLogo />
         </div>
         <LoginLinks />
       </div>
 
-      <section className="w-full bg-[#373435] text-white py-12 px-4">
+      <section className="w-full bg-primary-100 text-white p-[24px]">
         <div className="container mx-auto max-w-6xl">
-          <h1 className="text-4xl md:text-5xl font-bold mb-2">
-            Become Unos y Otros
-          </h1>
-          <p className="text-xl mb-8">
+          <h1 className="text-3xl font-medium mb-[4px]">Become Unos y Otros</h1>
+          <p className="text-base font-extralight mb-[20px]">
             Take control of your hobby with a managed cigar & humidor
             collection.
           </p>
           <div className="flex items-center">
             <Link
-              href="/login"
-              className="bg-white text-gray-800 font-bold py-3 px-8 rounded-full mr-4"
+              href="/register"
+              className="bg-white text-primary-100 font-semibold text-sm py-[10px] px-[15px] rounded-full"
             >
               Join Now
             </Link>
