+++ src/components/custom-icons/ApplicationLogo.tsx
@@ -1,0 +1,47 @@
+import React from "react";
+
+const ApplicationLogo = () => (
+  <>
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="67"
+      height="48"
+      viewBox="0 0 67 48"
+      fill="none"
+    >
+      <path
+        d="M47.0039 3.05176e-05L47.0039 33.0029L13.7573 29.6782C7.08283 29.0108 2.00005 23.3943 2.00005 16.6866C2.00005 10.0336 7.00245 4.44464 13.6147 3.70995L47.0039 3.05176e-05Z"
+        fill="#373435"
+      />
+      <mask id="path-2-inside-1_3387_126" fill="white">
+        <path d="M46.5041 33.0029C44.3371 33.0029 42.1914 32.576 40.1893 31.7468C38.1873 30.9175 36.3682 29.702 34.8359 28.1697C33.3036 26.6374 32.0881 24.8183 31.2588 22.8163C30.4295 20.8142 30.0027 18.6684 30.0027 16.5015C30.0027 14.3345 30.4295 12.1887 31.2588 10.1866C32.0881 8.18458 33.3036 6.36548 34.8359 4.83318C36.3682 3.30088 38.1873 2.0854 40.1893 1.25612C42.1914 0.426851 44.3371 2.95121e-05 46.5041 2.97963e-05L46.5041 1.94016C44.5919 1.94016 42.6984 2.31679 40.9318 3.04857C39.1651 3.78034 37.5599 4.85292 36.2077 6.20506C34.8556 7.5572 33.783 9.16243 33.0513 10.9291C32.3195 12.6957 31.9428 14.5892 31.9428 16.5015C31.9428 18.4137 32.3195 20.3072 33.0512 22.0738C33.783 23.8405 34.8556 25.4457 36.2077 26.7978C37.5599 28.15 39.1651 29.2226 40.9318 29.9543C42.6984 30.6861 44.5919 31.0627 46.5041 31.0627L46.5041 33.0029Z" />
+      </mask>
+      <path
+        d="M46.5041 33.0029C44.3371 33.0029 42.1914 32.576 40.1893 31.7468C38.1873 30.9175 36.3682 29.702 34.8359 28.1697C33.3036 26.6374 32.0881 24.8183 31.2588 22.8163C30.4295 20.8142 30.0027 18.6684 30.0027 16.5015C30.0027 14.3345 30.4295 12.1887 31.2588 10.1866C32.0881 8.18458 33.3036 6.36548 34.8359 4.83318C36.3682 3.30088 38.1873 2.0854 40.1893 1.25612C42.1914 0.426851 44.3371 2.95121e-05 46.5041 2.97963e-05L46.5041 1.94016C44.5919 1.94016 42.6984 2.31679 40.9318 3.04857C39.1651 3.78034 37.5599 4.85292 36.2077 6.20506C34.8556 7.5572 33.783 9.16243 33.0513 10.9291C32.3195 12.6957 31.9428 14.5892 31.9428 16.5015C31.9428 18.4137 32.3195 20.3072 33.0512 22.0738C33.783 23.8405 34.8556 25.4457 36.2077 26.7978C37.5599 28.15 39.1651 29.2226 40.9318 29.9543C42.6984 30.6861 44.5919 31.0627 46.5041 31.0627L46.5041 33.0029Z"
+        stroke="white"
+        strokeWidth="4.00034"
+        mask="url(#path-2-inside-1_3387_126)"
+      />
+      <mask id="path-3-inside-2_3387_126" fill="white">
+        <path d="M46.5041 33.0029C50.8806 33.0029 55.0778 31.2643 58.1724 28.1697C61.267 25.0751 63.0056 20.8779 63.0056 16.5015C63.0056 12.125 61.267 7.9278 58.1724 4.83319C55.0778 1.73857 50.8806 3.06484e-05 46.5041 2.97963e-05L46.5041 1.89544C50.3779 1.89544 54.093 3.43428 56.8321 6.17344C59.5713 8.9126 61.1101 12.6277 61.1101 16.5015C61.1101 20.3752 59.5713 24.0903 56.8321 26.8295C54.093 29.5686 50.3779 31.1075 46.5041 31.1075L46.5041 33.0029Z" />
+      </mask>
+      <path
+        d="M46.5041 33.0029C50.8806 33.0029 55.0778 31.2643 58.1724 28.1697C61.267 25.0751 63.0056 20.8779 63.0056 16.5015C63.0056 12.125 61.267 7.9278 58.1724 4.83319C55.0778 1.73857 50.8806 3.06484e-05 46.5041 2.97963e-05L46.5041 1.89544C50.3779 1.89544 54.093 3.43428 56.8321 6.17344C59.5713 8.9126 61.1101 12.6277 61.1101 16.5015C61.1101 20.3752 59.5713 24.0903 56.8321 26.8295C54.093 29.5686 50.3779 31.1075 46.5041 31.1075L46.5041 33.0029Z"
+        stroke="#373435"
+        strokeWidth="4.00034"
+        mask="url(#path-3-inside-2_3387_126)"
+      />
+      <path
+        d="M30.5026 17.0015L2.00017 17.0015"
+        stroke="white"
+        strokeWidth="2.00017"
+      />
+      <path
+        d="M3.68602 47.0935C3.12197 47.0935 2.66443 47.0005 2.3134 46.8145C1.96537 46.6284 1.71035 46.3554 1.54834 45.9954C1.38932 45.6354 1.30982 45.1928 1.30982 44.6678V40.4959H2.17389V44.7938C2.17389 45.3188 2.3044 45.7119 2.56542 45.9729C2.82645 46.2339 3.19998 46.3644 3.68602 46.3644C4.01305 46.3644 4.29057 46.3074 4.51859 46.1934C4.74961 46.0794 4.92513 45.9069 5.04514 45.6759C5.16815 45.4418 5.22965 45.1478 5.22965 44.7938V40.4959H6.06672V44.7533C6.06672 45.2543 5.98422 45.6789 5.8192 46.0269C5.65419 46.3749 5.39617 46.6404 5.04514 46.8235C4.69711 47.0035 4.24407 47.0935 3.68602 47.0935ZM7.61211 47.0035V40.4959H8.40868L11.847 45.5544V40.4959H12.6796V47.0035H11.937L8.44469 41.873V47.0035H7.61211ZM16.8506 47.0935C16.2715 47.0935 15.7705 46.963 15.3474 46.702C14.9244 46.4409 14.5974 46.0629 14.3664 45.5679C14.1383 45.0698 14.0243 44.4683 14.0243 43.7632C14.0243 43.0521 14.1398 42.4476 14.3709 41.9495C14.6019 41.4485 14.9289 41.066 15.3519 40.8019C15.778 40.5379 16.2775 40.4059 16.8506 40.4059C17.4266 40.4059 17.9247 40.5379 18.3447 40.8019C18.7647 41.063 19.0888 41.444 19.3168 41.945C19.5448 42.4431 19.6588 43.0491 19.6588 43.7632C19.6588 44.4683 19.5448 45.0698 19.3168 45.5679C19.0918 46.0629 18.7692 46.4409 18.3492 46.702C17.9292 46.963 17.4296 47.0935 16.8506 47.0935ZM16.8506 46.3689C17.2526 46.3689 17.5946 46.2774 17.8767 46.0944C18.1587 45.9114 18.3732 45.6279 18.5202 45.2438C18.6702 44.8598 18.7452 44.3693 18.7452 43.7722C18.7452 43.1661 18.6702 42.6681 18.5202 42.2781C18.3702 41.888 18.1542 41.6 17.8722 41.414C17.5901 41.225 17.2496 41.1305 16.8506 41.1305C16.4545 41.1305 16.114 41.225 15.829 41.414C15.544 41.603 15.3249 41.8925 15.1719 42.2826C15.0189 42.6726 14.9424 43.1691 14.9424 43.7722C14.9424 44.3693 15.0189 44.8598 15.1719 45.2438C15.3249 45.6279 15.544 45.9114 15.829 46.0944C16.114 46.2774 16.4545 46.3689 16.8506 46.3689ZM23.1866 47.0935C22.8686 47.0935 22.5671 47.056 22.282 46.981C21.997 46.903 21.7375 46.7875 21.5035 46.6344C21.2725 46.4814 21.0804 46.2924 20.9274 46.0674C20.7774 45.8394 20.6829 45.5754 20.6439 45.2753H21.517C21.571 45.5214 21.6775 45.7284 21.8365 45.8964C21.9985 46.0644 22.198 46.1904 22.4351 46.2744C22.6751 46.3584 22.9376 46.4004 23.2226 46.4004C23.4986 46.4004 23.7522 46.3569 23.9832 46.2699C24.2172 46.1829 24.4047 46.0569 24.5457 45.8919C24.6867 45.7239 24.7573 45.5229 24.7573 45.2888C24.7573 45.0308 24.6732 44.8238 24.5052 44.6678C24.3372 44.5088 24.0777 44.3813 23.7267 44.2852L22.3855 43.9387C22.0825 43.8637 21.814 43.7587 21.58 43.6237C21.349 43.4857 21.1674 43.3057 21.0354 43.0836C20.9034 42.8586 20.8374 42.5796 20.8374 42.2466C20.8374 41.8775 20.9394 41.5565 21.1434 41.2835C21.3475 41.0075 21.6235 40.7944 21.9715 40.6444C22.3225 40.4914 22.7186 40.4149 23.1596 40.4149C23.6697 40.4149 24.0987 40.5034 24.4467 40.6804C24.7978 40.8575 25.0633 41.09 25.2433 41.378C25.4233 41.666 25.5133 41.9735 25.5133 42.3006H24.6582C24.6312 42.0035 24.5442 41.7695 24.3972 41.5985C24.2502 41.4245 24.0672 41.3 23.8482 41.225C23.6292 41.147 23.3906 41.108 23.1326 41.108C22.6676 41.108 22.315 41.2145 22.075 41.4275C21.838 41.6375 21.7195 41.888 21.7195 42.1791C21.7195 42.4461 21.805 42.6561 21.976 42.8091C22.147 42.9621 22.399 43.0806 22.7321 43.1646L24.0372 43.5022C24.4242 43.5922 24.7317 43.7227 24.9598 43.8937C25.1908 44.0647 25.3558 44.2642 25.4548 44.4923C25.5568 44.7173 25.6078 44.9618 25.6078 45.2258C25.6078 45.5859 25.5073 45.9069 25.3063 46.1889C25.1053 46.4709 24.8233 46.693 24.4602 46.855C24.1002 47.014 23.6757 47.0935 23.1866 47.0935ZM30.667 47.0035V44.2312L28.4483 40.4959H29.3574L31.0856 43.4887L32.7867 40.4959H33.7003L31.4951 44.2357V47.0035H30.667ZM39.2031 47.0935C38.624 47.0935 38.123 46.963 37.6999 46.702C37.2769 46.4409 36.9499 46.0629 36.7189 45.5679C36.4908 45.0698 36.3768 44.4683 36.3768 43.7632C36.3768 43.0521 36.4923 42.4476 36.7234 41.9495C36.9544 41.4485 37.2814 41.066 37.7044 40.8019C38.1305 40.5379 38.63 40.4059 39.2031 40.4059C39.7791 40.4059 40.2772 40.5379 40.6972 40.8019C41.1172 41.063 41.4413 41.444 41.6693 41.945C41.8973 42.4431 42.0113 43.0491 42.0113 43.7632C42.0113 44.4683 41.8973 45.0698 41.6693 45.5679C41.4443 46.0629 41.1217 46.4409 40.7017 46.702C40.2817 46.963 39.7821 47.0935 39.2031 47.0935ZM39.2031 46.3689C39.6051 46.3689 39.9471 46.2774 40.2292 46.0944C40.5112 45.9114 40.7257 45.6279 40.8727 45.2438C41.0227 44.8598 41.0977 44.3693 41.0977 43.7722C41.0977 43.1661 41.0227 42.6681 40.8727 42.2781C40.7227 41.888 40.5067 41.6 40.2247 41.414C39.9426 41.225 39.6021 41.1305 39.2031 41.1305C38.807 41.1305 38.4665 41.225 38.1815 41.414C37.8965 41.603 37.6774 41.8925 37.5244 42.2826C37.3714 42.6726 37.2949 43.1691 37.2949 43.7722C37.2949 44.3693 37.3714 44.8598 37.5244 45.2438C37.6774 45.6279 37.8965 45.9114 38.1815 46.0944C38.4665 46.2774 38.807 46.3689 39.2031 46.3689ZM44.8973 47.0035V41.261H42.7956V40.4959H47.7191V41.261H45.7344V47.0035H44.8973ZM48.7924 47.0035V40.4959H51.3126C51.7807 40.4959 52.1722 40.5679 52.4872 40.7119C52.8023 40.853 53.0378 41.057 53.1938 41.324C53.3498 41.588 53.4278 41.9045 53.4278 42.2736C53.4278 42.6276 53.3648 42.9216 53.2388 43.1556C53.1158 43.3897 52.9583 43.5787 52.7663 43.7227C52.5742 43.8637 52.3747 43.9732 52.1677 44.0512L53.4458 47.0035H52.5637L51.3756 44.1997H49.625V47.0035H48.7924ZM49.625 43.5112H51.0426C51.5316 43.5112 51.9097 43.4122 52.1767 43.2142C52.4437 43.0131 52.5772 42.7161 52.5772 42.3231C52.5772 41.939 52.4572 41.6525 52.2172 41.4635C51.9802 41.2715 51.6547 41.1755 51.2406 41.1755H49.625V43.5112ZM57.2486 47.0935C56.6695 47.0935 56.1685 46.963 55.7454 46.702C55.3224 46.4409 54.9954 46.0629 54.7644 45.5679C54.5363 45.0698 54.4223 44.4683 54.4223 43.7632C54.4223 43.0521 54.5378 42.4476 54.7689 41.9495C54.9999 41.4485 55.3269 41.066 55.7499 40.8019C56.176 40.5379 56.6755 40.4059 57.2486 40.4059C57.8246 40.4059 58.3227 40.5379 58.7427 40.8019C59.1627 41.063 59.4868 41.444 59.7148 41.945C59.9428 42.4431 60.0568 43.0491 60.0568 43.7632C60.0568 44.4683 59.9428 45.0698 59.7148 45.5679C59.4898 46.0629 59.1672 46.4409 58.7472 46.702C58.3272 46.963 57.8276 47.0935 57.2486 47.0935ZM57.2486 46.3689C57.6506 46.3689 57.9926 46.2774 58.2747 46.0944C58.5567 45.9114 58.7712 45.6279 58.9182 45.2438C59.0682 44.8598 59.1432 44.3693 59.1432 43.7722C59.1432 43.1661 59.0682 42.6681 58.9182 42.2781C58.7682 41.888 58.5522 41.6 58.2702 41.414C57.9881 41.225 57.6476 41.1305 57.2486 41.1305C56.8525 41.1305 56.512 41.225 56.227 41.414C55.942 41.603 55.7229 41.8925 55.5699 42.2826C55.4169 42.6726 55.3404 43.1691 55.3404 43.7722C55.3404 44.3693 55.4169 44.8598 55.5699 45.2438C55.7229 45.6279 55.942 45.9114 56.227 46.0944C56.512 46.2774 56.8525 46.3689 57.2486 46.3689ZM63.5846 47.0935C63.2666 47.0935 62.9651 47.056 62.68 46.981C62.395 46.903 62.1355 46.7875 61.9015 46.6344C61.6705 46.4814 61.4784 46.2924 61.3254 46.0674C61.1754 45.8394 61.0809 45.5754 61.0419 45.2753H61.915C61.969 45.5214 62.0755 45.7284 62.2345 45.8964C62.3965 46.0644 62.596 46.1904 62.8331 46.2744C63.0731 46.3584 63.3356 46.4004 63.6206 46.4004C63.8966 46.4004 64.1502 46.3569 64.3812 46.2699C64.6152 46.1829 64.8027 46.0569 64.9437 45.8919C65.0847 45.7239 65.1553 45.5229 65.1553 45.2888C65.1553 45.0308 65.0712 44.8238 64.9032 44.6678C64.7352 44.5088 64.4757 44.3813 64.1247 44.2852L62.7836 43.9387C62.4805 43.8637 62.212 43.7587 61.978 43.6237C61.747 43.4857 61.5654 43.3057 61.4334 43.0836C61.3014 42.8586 61.2354 42.5796 61.2354 42.2466C61.2354 41.8775 61.3374 41.5565 61.5414 41.2835C61.7455 41.0075 62.0215 40.7944 62.3695 40.6444C62.7205 40.4914 63.1166 40.4149 63.5576 40.4149C64.0677 40.4149 64.4967 40.5034 64.8447 40.6804C65.1958 40.8575 65.4613 41.09 65.6413 41.378C65.8213 41.666 65.9113 41.9735 65.9113 42.3006H65.0562C65.0292 42.0035 64.9422 41.7695 64.7952 41.5985C64.6482 41.4245 64.4652 41.3 64.2462 41.225C64.0272 41.147 63.7886 41.108 63.5306 41.108C63.0656 41.108 62.713 41.2145 62.473 41.4275C62.236 41.6375 62.1175 41.888 62.1175 42.1791C62.1175 42.4461 62.203 42.6561 62.374 42.8091C62.545 42.9621 62.7971 43.0806 63.1301 43.1646L64.4352 43.5022C64.8222 43.5922 65.1298 43.7227 65.3578 43.8937C65.5888 44.0647 65.7538 44.2642 65.8528 44.4923C65.9548 44.7173 66.0058 44.9618 66.0058 45.2258C66.0058 45.5859 65.9053 45.9069 65.7043 46.1889C65.5033 46.4709 65.2213 46.693 64.8582 46.855C64.4982 47.014 64.0737 47.0935 63.5846 47.0935Z"
+        fill="#373435"
+      />
+    </svg>
+  </>
+);
+
+export default ApplicationLogo;
