+++ tailwind.config.js
@@ -1,5 +1,13 @@
 module.exports = {
   content: ["./src/**/*.{js,jsx,ts,tsx}"],
-  theme: {},
+  theme: {
+    extend: {
+      colors: {
+        primary: {
+          100: "#373435",
+        },
+      },
+    },
+  },
   plugins: [require("@tailwindcss/forms")],
 };
